+++ src/domain/song/repository.rs
@@ -9,5 +9,5 @@ pub trait SongRepository: Send + Sync {
         search: Vec<&str>,
         max_results: usize,
     ) -> Vec<Song>;
-    async fn search_by_db(&self, words: Vec<&str>, max_results: i64) -> Vec<Song>;
+    async fn search_by_db(&self, words: Vec<&str>, max_results: i32) -> Vec<Song>;
 }
